Basic fixes

+++ src/main/java/hu/montlikadani/ragemode/gameLogic/GameTimer.java
@@ -43,7 +43,7 @@ public class GameTimer {
     private void startCounting() {
         t = new Timer();
 
-        int totalTimerMillis = ((int) (((secondsRemaining * 1000) + 5000) / 10000)) * (10000);
+        int totalTimerMillis = (((secondsRemaining * 1000) + 5000) / 10000) * (10000);
         if(totalTimerMillis == 0)
             totalTimerMillis = 10000;
         final int timeMillisForLoop = totalTimerMillis;
